+++ lib/ControlBar.d.ts
@@ -1,8 +1,3 @@
-/// <reference types="react" />
-interface ControlBarProps {
-    onPressBack: () => void;
-    onFinishEditing: () => void;
-}
-declare function ControlBar(props: ControlBarProps): JSX.Element;
+declare function ControlBar(): JSX.Element;
 export { ControlBar };
 //# sourceMappingURL=ControlBar.d.ts.map