+++ lib/EditingWindow.d.ts
@@ -1,4 +1,3 @@
-/// <reference types="react" />
 declare function EditingWindow(): JSX.Element;
 export { EditingWindow };
 //# sourceMappingURL=EditingWindow.d.ts.map