+++ lib/OperationBar/Crop.d.ts
@@ -1,3 +1,2 @@
-/// <reference types="react" />
 export declare function Crop(): JSX.Element;
 //# sourceMappingURL=Crop.d.ts.map