+++ lib/OperationBar/Rotate.d.ts
@@ -1,3 +1,2 @@
-/// <reference types="react" />
 export declare function Rotate(): JSX.Element;
 //# sourceMappingURL=Rotate.d.ts.map