+++ lib/Processing.d.ts
@@ -1,4 +1,3 @@
-/// <reference types="react" />
 declare function Processing(): JSX.Element;
 export { Processing };
 //# sourceMappingURL=Processing.d.ts.map