+++ lib/components/IconButton.d.ts
@@ -1,4 +1,3 @@
-/// <reference types="react" />
 import { TouchableOpacityProps } from "react-native";
 import { IIconProps } from "./Icon";
 declare type IIconButtonProps = IIconProps & TouchableOpacityProps;