+++ lib/customHooks/usePerformCrop.d.ts
@@ -1,0 +1,2 @@
+export declare const usePerformCrop: () => () => Promise<void>;
+//# sourceMappingURL=usePerformCrop.d.ts.map